refactor: Move files into correct projects/namespaces
fixes #4
