+++ .editorconfig
@@ -1,5 +1,19 @@
-root = true
 # Remove the line below if you want to inherit .editorconfig settings from higher directories
+root = true
+
+# XML project files
+[*.{csproj,vbproj,vcxproj,vcxproj.filters,proj,projitems,shproj}]
+# Indentation and spacing
+indent_size = 2
+indent_style = space
+tab_width = 2
+
+# XML config files
+[*.{props,targets,ruleset,config,nuspec,resx,vsixmanifest,vsct}]
+# Indentation and spacing
+indent_size = 2
+indent_style = space
+tab_width = 2
 
 # C# files
 [*.cs]
@@ -83,9 +97,10 @@ csharp_style_var_when_type_is_apparent = true:warning
 csharp_style_expression_bodied_accessors = true:suggestion
 csharp_style_expression_bodied_constructors = false:suggestion
 csharp_style_expression_bodied_indexers = true:suggestion
-csharp_style_expression_bodied_lambdas = true:suggestion
+csharp_style_expression_bodied_lambdas = false:silent
+dotnet_diagnostic.RCS1021.severity = none
 csharp_style_expression_bodied_local_functions = when_on_single_line:suggestion
-csharp_style_expression_bodied_methods = false:suggestion
+csharp_style_expression_bodied_methods = true:silent
 csharp_style_expression_bodied_operators = when_on_single_line:suggestion
 csharp_style_expression_bodied_properties = true:suggestion
 
@@ -101,11 +116,11 @@ csharp_style_conditional_delegate_call = true:warning
 
 # Modifier preferences
 csharp_prefer_static_local_function = true:warning
-csharp_preferred_modifier_order = public,private,protected,internal,static,extern,new,virtual,abstract,sealed,override,readonly,unsafe,volatile,async
+csharp_preferred_modifier_order = public, private, protected, internal, static, extern, new, virtual, abstract, sealed, override, readonly, unsafe, volatile, async
 
 # Code-block preferences
-csharp_prefer_braces = false
 csharp_prefer_simple_using_statement = true
+csharp_prefer_braces = when_multiline:warning
 
 # Expression-level preferences
 csharp_prefer_simple_default_expression = true:warning
@@ -121,6 +136,9 @@ csharp_style_unused_value_expression_statement_preference = discard_variable:inf
 
 # 'using' directive preferences
 csharp_using_directive_placement = outside_namespace:warning
+
+# Convert to file-scoped namespace
+csharp_style_namespace_declarations = file_scoped:warning
 
 # New line preferences
 csharp_style_allow_blank_line_after_colon_in_constructor_initializer_experimental = false
@@ -202,52 +220,47 @@ dotnet_naming_rule.everything_but_fields_and_locals_should_be_pascal_case.style 
 
 dotnet_naming_symbols.interface.applicable_kinds = interface
 dotnet_naming_symbols.interface.applicable_accessibilities = *
-dotnet_naming_symbols.interface.required_modifiers = 
+dotnet_naming_symbols.interface.required_modifiers =
 
 dotnet_naming_symbols.parameters.applicable_kinds = parameter
 dotnet_naming_symbols.parameters.applicable_accessibilities = *
-dotnet_naming_symbols.parameters.required_modifiers = 
+dotnet_naming_symbols.parameters.required_modifiers =
 
 dotnet_naming_symbols.private_fields.applicable_kinds = field
 dotnet_naming_symbols.private_fields.applicable_accessibilities = private
-dotnet_naming_symbols.private_fields.required_modifiers = 
+dotnet_naming_symbols.private_fields.required_modifiers =
 
 dotnet_naming_symbols.type_parameters.applicable_kinds = type_parameter
 dotnet_naming_symbols.type_parameters.applicable_accessibilities = *
-dotnet_naming_symbols.type_parameters.required_modifiers = 
+dotnet_naming_symbols.type_parameters.required_modifiers =
 
 dotnet_naming_symbols.everything_but_fields_and_locals.applicable_kinds = namespace, property, event, class, struct, enum, delegate, method, local_function
 dotnet_naming_symbols.everything_but_fields_and_locals.applicable_accessibilities = *
-dotnet_naming_symbols.everything_but_fields_and_locals.required_modifiers = 
+dotnet_naming_symbols.everything_but_fields_and_locals.required_modifiers =
 
 # Naming styles
 
-dotnet_naming_style.pascal_case.required_prefix = 
-dotnet_naming_style.pascal_case.required_suffix = 
-dotnet_naming_style.pascal_case.word_separator = 
+dotnet_naming_style.pascal_case.required_prefix =
+dotnet_naming_style.pascal_case.required_suffix =
+dotnet_naming_style.pascal_case.word_separator =
 dotnet_naming_style.pascal_case.capitalization = pascal_case
 
 dotnet_naming_style.begins_with_i.required_prefix = I
-dotnet_naming_style.begins_with_i.required_suffix = 
-dotnet_naming_style.begins_with_i.word_separator = 
+dotnet_naming_style.begins_with_i.required_suffix =
+dotnet_naming_style.begins_with_i.word_separator =
 dotnet_naming_style.begins_with_i.capitalization = pascal_case
 
-dotnet_naming_style.parameters_camel_case.required_prefix = 
-dotnet_naming_style.parameters_camel_case.required_suffix = 
-dotnet_naming_style.parameters_camel_case.word_separator = 
+dotnet_naming_style.parameters_camel_case.required_prefix =
+dotnet_naming_style.parameters_camel_case.required_suffix =
+dotnet_naming_style.parameters_camel_case.word_separator =
 dotnet_naming_style.parameters_camel_case.capitalization = camel_case
 
 dotnet_naming_style.private_field_style.required_prefix = _
-dotnet_naming_style.private_field_style.required_suffix = 
-dotnet_naming_style.private_field_style.word_separator = 
+dotnet_naming_style.private_field_style.required_suffix =
+dotnet_naming_style.private_field_style.word_separator =
 dotnet_naming_style.private_field_style.capitalization = camel_case
 
 dotnet_naming_style.generic_parameter_style.required_prefix = T
-dotnet_naming_style.generic_parameter_style.required_suffix = 
-dotnet_naming_style.generic_parameter_style.word_separator = 
+dotnet_naming_style.generic_parameter_style.required_suffix =
+dotnet_naming_style.generic_parameter_style.word_separator =
 dotnet_naming_style.generic_parameter_style.capitalization = pascal_case
-
-# ReSharper inspection severities
-resharper_arrange_accessor_owner_body_highlighting = hint
-resharper_arrange_constructor_or_destructor_body_highlighting = hint
-resharper_arrange_method_or_operator_body_highlighting = hint
