--- Commands/EditBankAccount/EditBankAccountCommand.cs
@@ -1,7 +1,0 @@
-namespace Commands.EditBankAccount;
-
-public record EditBankAccountCommand(
-       Guid Id,
-       string Name,
-       string AccountNumber
-    ) : IRequest<Guid>;
