+++ Commands/Handlers/Transaction/AddCreditTransaction/AddCreditTransactionHandler.cs
@@ -3,8 +3,6 @@ using Commands.Handlers.Transaction.AddAttachments;
 
 using Domain;
 
-using MediatR;
-
 using Persistence.Repositories;
 
 namespace Commands.Handlers.Transaction.AddCreditTransaction;
