+++ Commands/Handlers/Transaction/AddDebitTransaction/AddDebitTransactionCommand.cs
@@ -2,8 +2,6 @@
 
 using FluentValidation;
 
-using MediatR;
-
 namespace Commands.Handlers.Transaction.AddDebitTransaction;
 
 public record AddDebitTransactionCommand(
