+++ Domain/Domain.csproj
@@ -3,5 +3,8 @@
   <ItemGroup>
     <ProjectReference Include="..\Types\Types.csproj" />
   </ItemGroup>
-
+  <ItemGroup>
+    <PackageReference Include="AutoMapper" Version="11.0.1" />
+    <PackageReference Include="Azure.Storage.Blobs" Version="12.12.0" />
+  </ItemGroup>
 </Project>
