+++ Types/Attachment.cs
@@ -1,3 +1,3 @@
-﻿namespace Commands.Handlers.Transaction;
+﻿namespace Types;
 
 public record Attachment(string Name, string ContentType, string BlobUri, byte[] Bytes);
