+++ Web/Pages/BankAccounts/NewBankAccount.razor
@@ -1,7 +1,7 @@
 @page "/bankaccounts/new"
 @attribute [Authorize]
 
-@using Commands.Handlers.BankAccount
+@using Commands.Handlers.BankAccount.AddBankAccount
 @inject IMediator _mediator
 @inject IMapper _mapper
 @inject ISnackbar Snackbar
