+++ Web/Pages/Members/DeleteMemberDialog.razor
@@ -1,4 +1,5 @@
 ﻿@using Commands
+@using Commands.Handlers.Member.DeleteMember
 @using Queries.Members.ViewModels
 @using MediatR;
