+++ Web/Pages/Members/EditMember.razor
@@ -2,6 +2,7 @@
 @attribute [Authorize]
 
 @using Commands
+@using Commands.Handlers.Member.EditMember
 @using MediatR
 @using AutoMapper
 @using Queries.Members.Handlers
