+++ Web/Pages/Members/NewMember.razor
@@ -2,6 +2,7 @@
 @attribute [Authorize]
 
 @using Commands
+@using Commands.Handlers.Member.AddMember
 @using MediatR
 @using AutoMapper
 @inject IMediator _mediator
