+++ Web/Pages/Transactions/AddBankAccountDialog.razor
@@ -1,5 +1,5 @@
 ﻿@using Web.Models
-@using Commands.Handlers.BankAccount
+@using Commands.Handlers.BankAccount.AddBankAccount
 @using MediatR
 
 @inject IMediator mediator;
